requirements.txt
joblib==0.11
numpy==1.12.1
tensorflow-gpu==0.12.0
keras==1.2.2